+++ CHANGES.md
@@ -6,7 +6,7 @@
 
 * Fix connection get closed too early
 
-* Fix steraming response handling for http/2
+* Fix streaming response handling for http/2
 
 * Better sleep on error support