+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changes
 
+## 0.4.8 (2018-03-xx)
+
+* Do not log WouldBlock error from socket accept
+
 ## 0.4.7 (2018-03-11)
 
 * Fix panic on unknown content encoding